backend/requirements.txt
fastapi>=0.104
uvicorn>=0.24
anthropic>=0.40
pdfplumber>=0.10
python-docx>=1.1
python-multipart>=0.0.6
